+++ src/SqlArtisan/Internal/SqlBuilder/Select/ISelectBuilderJoin.cs
@@ -1,6 +1,6 @@
 ﻿namespace SqlArtisan.Internal;
 
-public interface ISelectBuilderJoin : ISqlBuilder
+public interface ISelectBuilderJoin : ISqlBuilder, IForUpdate
 {
     // Subsequent SQL is the same as the FROM clause.
     ISelectBuilderFrom On(SqlCondition condition);