+++ src/SqlArtisan/Internal/SqlBuilder/Select/ISelectBuilderWhere.cs
@@ -1,6 +1,6 @@
 ﻿namespace SqlArtisan.Internal;
 
-public interface ISelectBuilderWhere : ISqlBuilder, ISetOperator, ISubquery
+public interface ISelectBuilderWhere : ISqlBuilder, ISetOperator, IForUpdate, ISubquery
 {
     ISelectBuilderGroupBy GroupBy(params object[] groupByItems);