+++ src/SqlArtisan/Internal/SqlPart/Clause/ForUpdate/NowaitBehavior.cs
@@ -1,0 +1,7 @@
+﻿namespace SqlArtisan.Internal;
+
+public sealed class NowaitBehavior : LockBehaviorBase
+{
+    internal override void Format(SqlBuildingBuffer buffer) =>
+        buffer.Append(Keywords.Nowait);
+}